requirements.txt
fastapi==0.82.0
typing==3.7.4.3
uvicorn==0.18.3
opencv-python==4.6.0.66
numpy==1.21.2
torch==1.12.1
torchvision==0.13.1
Pillow==9.2.0
PyMuPDF==1.21.1
setuptools==63.2.0
einops==0.2.0
gdown==4.4.0
PyYAML==5.3.1
pandas==1.4.4
matplotlib==3.5.3
seaborn==0.12.0
scipy==1.9.1
python-multipart==0.0.5
random2==1.0.1
pytesseract==0.3.10